mm: Add extra free kbytes tunable
Add a userspace visible knob to tell the VM to keep an extra amount of
memory free by increasing the gap between each zone's min and low
watermarks.

This is useful for realtime applications that use syscalls and have a
bound on the number of allocations that happen in any short time period.
In these applications extra_free_kbytes would be left at an amount equal
to or larger than the maximum number of allocations that happen in any
burst.

It may also be useful to reduce the memory use of virtual machines
(temporarily?) in a way that does not cause memory fragmentation like
ballooning does.

[ccross]
Revived for use on old kernels where no other solution exists. The tunable
will be removed on kernels that do better at avoiding direct reclaim.

Change-Id: I765a42be8e964bfd3e2886d1ca85a29d60c3bb3e
Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Colin Cross <ccross@android.com>
Git-commit: 92189d47f66c67e5fd92eafaa287e153197a454f
Git-repo: https://android.googlesource.com/kernel/common/
[bhargavuln@codeaurora.org: resolve trivial merge conflicts]
Signed-off-by: Bhargav Upperla <bhargavuln@codeaurora.org>