kernel: timeconst: Rewrite in BC
BC is a standard tool for multi-precision arithmetic. We switched to Perl,
because akpm reported a hard-to-reproduce build hang, which was very odd,
because affected and unaffected machines were all running the same version
of GNU BC.

Unfortunately switching to Perl required a really ugly "canning" mechanism
to support pre 5.8 Perl installations lacking the Math::BigInt module.

It was recently pointed out to me that some very old versions of GNU make
had problems with pipes in subshells, which was indeed the construct used
in the Makefile rules in that version of the patch; Perl didn't need it so
switching to Perl fixed the problem for unrelated reasons.  With the
problem (hopefully) root-caused, we can switch back to BC and do the
arbitrary-precision arithmetic naturally.

Change-Id: I8450a919c2d27b6c18561621c0a48a762e46a22d
Signed-off-by: H. Peter Anvin <hpa@zytor.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Acked-by: Sam Ravnborg <sam@ravnborg.org>
Signed-off-by: Michal Marek <mmarek@suse.cz>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>