msm: SSR: Remove useless warning
This is a warning so that developers know to add more restart order lists
to SSR when a new chip is added. This is mostly irrelevant now because we
assume either entire SoC restart on SSR or independent restart on SSR, not
group restarts. Remove this warning as it is mostly a reminder that nobody
is listening to.

Change-Id: Icbf955cb18395d8d5d086b2167c5c329588b9256
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>