fs: sync: Allow to temporary disable a part of file synchronization
Leaving a machine without a reliable way to discard freshly changed data
is a crazy idea just on its own, but it is also an easy way to get a huge
file system performance burst. Of course, IT MUST NOT BE USED IN ANY CASE
as it creates a tremendous hole in FS subsytem which can lead to a giant
data loss, but if you are sure the chance of a sudden system or power
crash is low and you want to significantly improve storage performance for
a short period of time, then you can temporarly avoid a part of file
synchronization via a new module variable.

I'll repeat that this must not be ever used for a long period of time
(and for short too, though).

Change-Id: Ie7236953887affa6606c85250ba3290d3f81627a
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>